Bitwise reading from file workin
The implementation right now is so that
the piece is filled from most significant
bit "backwards". This is mainly due to
easier visualization of such structure.
But, a fill starting from least significant
bit could decrease calculations and improve
overall performance. This detail must be
decided before full implementation of the
solver.