The input from least significant as uint16_t works
Previously the bitwise implementation
was storing data from the most to least
significant bit, which would have caused
silly issues in the solving stage due to
weird bit alignment. Now that the bits are
stored from "right to left", i.e. from
least to most significant bit, the casting
to uint64_t on the board will be painless
and the function fast.