Block contact counter fixed from file.c
Previously the file format checker wasn't
functioning when the count_contact -function
was separated as it's own function. The reason was
that the indexing was applied twice: the buffer
was referenced to the function by index, and then
read with that index again inside the function.
Now (using pointer arithmetic) the function works.