Fixed buffer overflow in solve_driver
The memset() called for char array 'solution' was writing
an extra 16 bytes of dot ('.') outside the memory space of
'solution', due to a remnant of a previous string (i.e. null
terminated) solution. This bug does not affect the functions
workings, but is an overflow nonetheless and should be fixed
by default.